LayoutTests/imported/w3c/web-platform-tests/cssom/CSSNamespaceRule-expected.txt


FAIL CSSOM - CSSNamespaceRule interface assert_equals: CSSNamespaceRule cssText attribute expected "@namespace svg url(\"http://servo\");" but got "@namespace svg url('http://servo');"
